+++ gwcelery/conf/playground.py
@@ -1,6 +1,9 @@
 """Application configuration for ``gracedb-playground.ligo.org``."""
 
-from .test import *  # noqa: F401, F403
+from . import *  # noqa: F401, F403
+
+lvalert_host = 'lvalert-playground.cgca.uwm.edu'
+"""LVAlert host."""
 
 gracedb_host = 'gracedb-playground.ligo.org'
 """GraceDb host."""